Refactor test ServerTest structure
Split ServerTest per feature test classes
